requirements.txt
Flask==2.3.2
transformers==4.41.1
torch>=2.0.0
PyMuPDF==1.23.25
python-docx==1.1.0
<<<<<<< HEAD
pdfplumber==0.10.3
=======
pdfplumber==0.10.3
langchain==0.1.16
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyMuPDF==1.23.7
transformers==4.40.0
langchain-google-genai==0.0.6
google-generativeai==0.5.4
>>>>>>> 72f408b (final)
